+++ lib/core/api/index.js
@@ -1,0 +1,7 @@
+exports.bucket = require('./bucket.js');
+exports.createFile = require('./create-file.js');
+exports.getFile = require('./get-file.js');
+exports.createFileToken = require('./create-file-token.js');
+exports.getFilePointers = require('./get-file-pointers');
+exports.initStore = require('./init-store.js');
+exports.resolveFileFromPointers = require('./resolve-file-from-pointers.js');
