+++ lib/core/api/resolve-file-from-pointers.js
@@ -1,0 +1,48 @@
+/**
+  * @module storj/file
+  * @license LGPL-3.0
+  */
+
+'use strict';
+
+var assert = require('assert');
+var calculateFileId = require('storj-lib/lib/utils.js').calculateFileId;
+var KeyIV = require('storj-lib/lib/crypto-tools/deterministic-key-iv.js');
+var EncryptStream = require('storj-lib/lib/crypto-tools/encrypt-stream.js');
+//var Client = require('storj-lib/lib/bridge-client');
+var crypto = require('crypto');
+var storjutils = require('storj-lib/lib/utils.js');
+var sha256 = require('storj-lib/lib/utils').sha256;
+var store = require('memory-blob-store');
+var util = require('util');
+var crypt = require('crypto');
+
+
+
+module.exports = function(self) {
+  // TODO: implement
+  /**
+    * Conveinence function for emitting an error. If there is a registered
+    * listener on the File, then we will emit there, otherwise we will attempt to
+    * emit on the client. If there is not a client, we throw the error. Client may
+    * throw an unhandled error event exception if no listeners are registered.
+    * @private
+   */
+  var _error = function (e) {
+    // if(this.listenerCount('error') > 0) {
+    //   return this.emit('error', e);
+    // }
+    // if(self._client) {
+    //   return self._client.emit('error', e);
+    // }
+    throw e;
+  };
+
+  /**
+    * Given a list of pointers, this function will reconstruct the file
+    * @private
+  */
+  return function(pointers, opts, cb){
+    self._client.resolveFileFromPointers(pointers, opts, cb);
+  }
+}
