+++ lib/core/index.js
@@ -1,7 +1,5 @@
 'use strict'
 
-var file = require('./file.js');
-var bucket = require('./bucket.js');
 //var Client = require('./Client.js');
 // Use bridge client for now
 var Client = require('storj-lib/lib/bridge-client');
@@ -10,6 +8,7 @@ var util = require('util');
 var EventEmitter = require('events').EventEmitter;
 var assert = require('assert');
 var store = require('memory-blob-store');
+var api = require('./api');
 
 /**
  * Creates a new Storj object for interacting with the Storj network
@@ -39,14 +38,19 @@ function Storj(opts) {
       basicAuth: this._basicAuth,
       store: this._store
     });
-  
+
+  // TODO: Bind this context to new File()
+
   // core components 
-  this.file = file(this);
-  this.bucket = bucket(this);
+  //this.file = file(this);
+  this.bucket = api.bucket(this);
   this.createBucket = this.bucket.createBucket; // alias for createBucket
   this.deleteBucket = this.bucket.deleteBucket; // alias for deleteBucket
-  this.createFile = this.file.createFile; // alias for createFile
-  this.getFile = this.file.getFile; // alias for getFile
+  this.createFile = api.createFile(this);
+  this.getFile = api.getFile(this);
+  this.createFileToken = api.createFileToken(this);
+  this.getFilePointers = api.getFilePointers(this);
+  this.initStore = api.initStore(this);
 
   if(opts.key) {
     /* TODO */
